refactor: 신고 관련 모달 컴포넌트 파일 위치 수정
[#363]
