+++ frontend/src/components/ui/DetailedStationInfo/reports/ChargerReportConfirmation.tsx
