+++ frontend/src/components/ui/DetailedStationInfo/reports/StationReportConfirmation.tsx
@@ -15,8 +15,8 @@ import TextField from '@common/TextField';
 
 import StationInformation from '@ui/DetailedStationInfo/StationInformation';
 
-import type { StationDetails } from '../../../types';
-import type { ChargerDetails } from '../../../types/chargers';
+import type { StationDetails } from '../../../../types';
+import type { ChargerDetails } from '../../../../types/chargers';
 
 interface StationReportConfirmationProps {
   station: StationDetails;
